Meghna, could you review?
New package: jcradaptor.  This adapts a jcr-implementation as a connector.spi implementation.  This could be used by customers or 3rd-party developers to adapt to existing jcr implementations.  For us, we will use it with the MockRepository and MockJcrRepository for writing stand-alone unit tests.  We could also use it to write test connectors to open-source jcr-compliant repositories, such as alfresco and jackrabbit.

The hardest part is implementing resumeTraversal().  Suppose the last repository has multiple documents that have the same date stamp.  This can happen through bulk-load or some kinds of backup-restore.  If the query for "documents-modified-since" uses a > (strictly-greater-than) operator, there's a danger that some documents might be missed.  If it uses a >= (greater-than-or-equal) operator, there's a danger that no progress will be made, it might process the same documents over and over.

Thus (using SQL here for clarity), it is important to do a query like this:
   select ... from documents where lastModified >= date-constant order by lastModified, documentid
That way, the >= makes sure you don't miss anything while the second order-by clause allows you to skip up to the last one you processed, avoiding repeatedly processing the same documents and the danger of making no progress.

This is reflected primarily in SpiQueryTraversalManagerFromJcr.java, which generates the query (xpath, not sql, sadly), and advances through the result to the right point.  There is also complication in SpiResultSetFromJcr.java, where we need to compensate for the fact the java Iterators do not permit you to take something, then push it back.

The other classes are quite straightforward mappings from jcr classes to nearly identical spi classes.

In addition, test data files are in this check-in - I think they should have gone in earlier.  Also, a useful end-to-end traversal test.