Danny, please review.
Added authn and authz to the mock repository, to the mock jcr repository, and to the jcradaptor.  This will allow end-to-end testing of Authentication and Authorization managers.  

Authorization in the mock repository is implemented through a property called "acl", with repeating String values.  If a document has no acl, it is public.  If it has an acl, then it is readable only by the users listed in the acl.

Authorization is implemented through a special document in the repository with uuid "users".  Its acl is a list of all users that have accounts in the repository.  Their passwords are the same as their names and can not be changed.