Danny, please review.
Rename Configurer as ConnectorType and Repository as Connector

Three SPI files are affected: Configurer, Repository and SimpleConfigurer.  The other files in this checkin were automatically changed by eclipse when I renamed those three.  

I verified the comments in those three files but there may be other files where comments should be changed to match this naming.  Please take a look at files you're currently workiong on and fix the comments as we go.  There will probably be a few more changes to follow on this.  BTW, I tried to use svn rename to preserve history.  We'll see whether that worked...