Meghna, please review.
I did a little file renaming to help me understand what's going on.

UrlConn and friends are now FeedConnection.  FeedUrlConn.java (the non-mock one) is now GsaFeedConnection.  MockUrlCOnn also moved to the pusher package - the mock package is for mock objects that have no corresponding interface in the java directory that they implement - if they implement an interface in java, then they should be in that package.