Begin refactor of traversal spi.
** Connector developers should not take this checkin. **

The repo will be in flux for a few checkins.  This checkin compiles and all unit tests pass, but it is not in a logically consistent state.
This checkin:

Moves obsolete parts of the spi to spi/old temporarily.  This is so that everything can still compile and run, while the new spi is introduced gradually.  

Introduces the new spi classes and interfaces.  There are no other changes anticipated to the spi as part of this series of checkins.

However: the connector manager still uses the old spi, so all of the new spi classes and interfaces are dead code.  The next few check-ins will gradually shift the connector manager to the new spi, and finally, the spi/old directory will become dead and will be removed.