This series of changes allows the Connector Manager's connector logs and feed logs to be fetched via a servlet interface.  The servlet allows the user to list available log files, fetch individual log files, and fetch a ZIP archive of all the log files.
Usage:
-----
To list the available connector log files:
  http://[cm_host_addr]/connector_manager/getConnectorLogs

To view an individual connector log file:
  http://[cm_host_addr]/connector_manager/getConnectorLogs/[log_file_name]
where [log_file_name] is the name of one log files returned by the list.
For instance, 'google-connectors.otex0.log'.  As a convenience, the log 
name may be simply the log file generation number, '0' in the above example,
and it gets automatically expanded.  Also, as a convenience, the servlet
name 'getConnectorLog' may be used (note the singular).

To retrieve a ZIP archive of all the connector log files:
  http://[cm_host_addr]/connector_manager/getConnectorLogs/*
or
  http://[cm_host_addr]/connector_manager/getConnectorLogs/ALL

To list the available feed log files:
  http://[cm_host_addr]/connector_manager/getFeedLogs

To view an individual feed log file:
  http://[cm_host_addr]/connector_manager/getFeedLogs/[log_file_name]
where [log_file_name] is the name of one log files returned by the list.
For instance, 'google-connectors.feed0.log'.  As a convenience, the log 
name may be simply the log file generation number, '0' in the above example,
and it gets automatically expanded.  Also, as a convenience, the servlet
name 'getFeedLog' may be used (note the singular).

To retrieve a ZIP archive of all the feed log files:
  http://[cm_host_addr]/connector_manager/getFeedLogs/*
or
  http://[cm_host_addr]/connector_manager/getFeedLogs/ALL


Upgrading existing Connector Manager installations:
--------------------------------------------------
Note that since this change modifies both the Connector Manager 
applicationContext.xml and its web.xml, upgrading existing CM
installations is more difficult then simply dropping in a new
jar file and restarting Tomcat.  These XML files must be updated
in existing deployments as well.  This might be as simple as 
replacing the existing deployed version with the new version,
however beware of overwriting customer modifications to these
files.  In my case, I also had to update my Connector Manager
applicationContext.properties file to include the feedLoggingLevel
property.


Change Log:
----------

* projects/connector-manager/source/java/com/google/enterprise/connector/servlet/GetConnectorLogs.java
  - New class that implements the log fetching capability servlet.

* projects/connector-manager/source/java/com/google/enterprise/connector/servlet/ServletUtil.java
  - Added additional XML tags and MIME types used when fetching logs.

* projects/connector-manager/source/java/com/google/enterprise/connector/pusher/FeedFileHandler.java
  - A new trivial subclass of java.util.logging.FileHandler that allows the
    retrieval of the pattern string used to generate the log file names.

* projects/connector-manager/etc/applicationContext.xml
  - Changed the FeedHandler bean instantiator to construct a FeedFileHandler
    object, rather than a java.util.logging.FileHandler object.

* projects/connector-manager/source/webdocs/web.xml
  - Added the servlet interfaces for retrieving the logs.