This set of changes addresses Connector Manager Issue 112 - Use PropertyPlaceholderConfigurer for feedLoggingLevel property
This is a couple usability and performance improvements,
mostly dealing with logging.

When a customer was asked to turn off feed logging,
he commented out the feedLoggingLevel property in
the CM applicationContext.properties rather than
setting the property value to "OFF".  This caused
the Connector Manager to fail to start up with a
failure to resolve a property placeholder.

One easy solution was to provide a default value for
the property in the PropertyPlaceholderConfigurer
bean of the Connector Manager's applicationContext.xml
file.

This same customer complained about poor feed
performance when teedFeedFile was enabled.  Since
the teedFeedFile tends to get volumes of data, I
increased the BufferedOutputStream buffer size from
the default 512 bytes to 32768 bytes.


Change Log:
----------
M  connector-manager/source/java/com/google/enterprise/connector/pusher/DocPusher.java
   - Increase the teedFeedFile BufferedOutputStream
     buffer size to improve feed logging performance.

M  connector-manager/etc/applicationContext.xml
   - Provide default values for feedLoggingLevel and
     gsa.feed.port properties in the
     PropertyPlaceholderConfigurer.