Implement synchronization between
  o operations on running traversals such as start, cancel, save 
    completion information
  o operations that reconfigure a connector such as remove connector,
    set configuration parameters.

using the ConnectorCoordinator rather than earlier mechanisms. This 
change is intended to make the code simpler to understand because
all the coordination is in one place and it all depends on the monitor
for a single object (the ConnectorCoordinatorImpl).

This change removes
  o Generational store and InstanceInfo/ConnectorInterfaces logic for
    managing the GenerationalStore
  o TraversalSecheduler tracking of running batches and code to handle
    races between scheduling and removing a batch.
  o ConnectorInterfaces/QueryTraverser asyncronous invalidation/cancel
    of query traversers.

After this change the traversal scheduler still
  o reads a schedule, releases the ConnectorCoordinator monitor an
    decides if a traversal should run
  o obtains the ConnectorCoordinatorMonitor and runs the batch.

It may be advisable in a future change to preform both of these
operations while holding the ConnectorCoordinator monitor so the
schedule cannot be changed between.