Added a SecureDocument utility class to simplify the constructor of ACLs and documents with ACLs.
I have put it into the spi package. If we copy the other utility
classes, especially the Simple classes to the util package and
deprecate the ones in the spi package, then we can move this to the
util package at the same time.

Other changes:

* Renamed InheritanceType to AclInheritanceType for consistency.
* Renamed Scope and Access to AclScope and AclAccess for clarity, and
  moved them to SpiConstants so that SecureDocument can use them.
* Extracted a createSpiProperties method in ConnectorTestUtils.

Code review: http://codereview.appspot.com/5616049/