Fix issue 6734754: GSA cannot handle null Schedule.
The GSA, while claiming to support null or empty schedules,
actually does not.  This change avoids sending a null
schedules to the GSA from Mangager.getConnectorStatus[es].
If a connector does not have a schedule, getConnectorStatus
will supply a schedule with default load, retryDelay, and
interval, but disabled.

To assist with this, the Schedule class was enhanced with
a factory method for optionally creating a schedule (if
the serialized schedule is not null or empty). The no-args
constructor returns a disabled schedule with default load,
retry delay, and time intervals.  And finally, a static
toString(Schedule) method was added that returns a 
serialized disabled default Schedule if the supplied
Schedule is null.

Code Review:
http://codereview.appspot.com/7138052