Deprecate the document store and remove the implementation.
The document store was never used or exposed in a release. It was
intended as an audit log and eventually as a persistent store for
diffing or other highly stateful implementations. We are highly
unlikely to ever revive this code, and the threading issues with H2
are the coup de grace.

I kept the internal DocumentStore interface (but deprecated it),
because it is used in the QueryTraverser constructor and
DocPusher.take method. These are used by connectors in their tests, so
I decided not to break them immediately. I deprecated the existing
constructor and method, and added new overloads without the
DocumentStore parameter. I did not give DocumentAcceptorImpl the same
treatment, since it has a public interface and is not similarly used
by the Google connectors at least.

This changes causes a test failure in QueryTraverserTest. I'm going to
take that up in a separate CL, because the failure s related to a
series of issues that are not related to this already large change.

Code review: https://codereview.appspot.com/12829045/