Remove GData JAR files and xercesImpl-2.8.1.jar. These were added in r2952 and r2954 to support SocialCollectionHandler, but that functionality was removed in r3329 as part of the broader change.
Connector Manager no longer uses these files. The SharePoint connector
has its own copy of xercesImpl-2.8.1.jar. The gdata-core-1.0.jar file
was added to the file system connector as part of the broader change.
The Notes connector has its own copy of the GData JAR files.

This is part of removing the GSA Admin API so that we can upgrade
Guava. The GSA Admin API depends on the GData library, which in turn
depends on Guava version 10 or earlier.

Code review: https://codereview.appspot.com/125460043/